Added multiqc plugins configuration
