+++ assets/multiqc_plugins/multiqc_aladdin_viralrecon/__init__.py
@@ -3,5 +3,5 @@
 from pkg_resources import get_distribution
 from multiqc.utils import config
 
-__version__ = get_distribution("multiqc_aladdin_rnaseq").version
-config.multiqc_aladdin_rnaseq_version = __version__
+__version__ = get_distribution("multiqc_aladdin_viralrecon").version
+config.multiqc_aladdin_viralrecon_version = __version__
