+++ assets/multiqc_plugins/multiqc_aladdin_viralrecon/templates/aladdin/__init__.py
@@ -1,0 +1,6 @@
+import os
+
+template_parent = 'default'
+
+template_dir = os.path.dirname(__file__)
+base_fn = 'base.html'
