--- assets/multiqc_plugins/multiqc_aladdin_viralrecon/templates/aladdin/assets/__init__.py
@@ -1,6 +1,0 @@
-import os
-
-template_parent = 'default'
-
-template_dir = os.path.dirname(__file__)
-base_fn = 'base.html'
