+++ assets/multiqc_plugins/multiqc_aladdin_viralrecon/templates/aladdin/content.html
